@@ -3,10 +3,10 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 31
     }
     buildTypes {
         release {
@@ -24,7 +24,6 @@ dependencies {
     implementation "com.google.code.gson:gson:$gsonConverter"
     implementation "com.squareup.retrofit2:retrofit:$retrofit"
     implementation "com.squareup.retrofit2:converter-gson:$gson"
-    implementation "com.squareup.okhttp3:okhttp:$okhttp"
     implementation "com.squareup.okio:okio:$okio"
     implementation "com.squareup.okhttp3:logging-interceptor:$interceptor"
     implementation project(path: ':domain')
@@ -34,7 +33,7 @@ dependencies {
 sourceCompatibility = "7"
 targetCompatibility = "7"
 buildscript {
-    ext.kotlin_version = '1.3.50'
+    ext.kotlin_version = '1.6.10'
     repositories {
         mavenCentral()
     }
