@@ -1,8 +1,7 @@
 package com.past3.ketroapp.di
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
-import java.lang.IllegalArgumentException
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import javax.inject.Inject
 import javax.inject.Provider
 import javax.inject.Singleton
@@ -10,13 +9,13 @@ import javax.inject.Singleton
 @Singleton
 @Suppress("UNCHECKED_CAST")
 class ViewModelFactory @Inject constructor(
-        private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
+    private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
 ) : ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         val creator = creators[modelClass]
-                ?: creators.asIterable().firstOrNull { modelClass.isAssignableFrom(it.key) }?.value
-                ?: throw IllegalArgumentException("Unknown BaseViewModel class $modelClass")
+            ?: creators.asIterable().firstOrNull { modelClass.isAssignableFrom(it.key) }?.value
+            ?: throw IllegalArgumentException("Unknown BaseViewModel class $modelClass")
 
         return try {
             creator.get() as T
