@@ -1,7 +1,7 @@
 package com.past3.ketro.api
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import com.past3.ketro.kcore.model.StatusCode
 import com.past3.ketro.kcore.model.Wrapper
 import retrofit2.Call
