@@ -1,11 +1,11 @@
 package com.past3.ketro.presentation
 
-import android.support.v7.recyclerview.extensions.ListAdapter
-import android.support.v7.util.DiffUtil
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
 import com.past3.ketro.domain.entities.Items
 import kotlinx.android.synthetic.main.list_item.view.*
 
